finishing up backend to where it was
