--- data/exported/task.txt
@@ -1,1 +1,0 @@
-Test.java*main
