+++ start_webserver.bat
@@ -1,3 +1,2 @@
-cd ./src/backend
 node server.js
 pause
